+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### :wrench: Changed
 
 - Generate at least one non-NULL character in path parameters. [#2790](https://github.com/schemathesis/schemathesis/issues/2790)  
+- Improved failure messages for `unsupported_method` and `missing_required_header` checks.
 
 ## [4.0.0-alpha.12](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.11...v4.0.0-alpha.12) - 2025-06-02